Conflicts:
	bin/common/Language/en-GB.yml
	bin/common/Language/en-US.yml
	bin/standard/xcom2/extraSprites.rul
	cmake/modules/FindSDL2.cmake
	src/Basescape/CraftSoldiersState.h
	src/Basescape/ManufactureInfoState.cpp
	src/Basescape/PurchaseState.cpp
	src/Battlescape/BattlescapeState.cpp
	src/Battlescape/BattlescapeState.h
	src/Battlescape/Camera.cpp
	src/Battlescape/Inventory.cpp
	src/Battlescape/InventoryState.cpp
	src/Battlescape/MiniMapView.cpp
	src/Engine/FlcPlayer.cpp
	src/Engine/OpenGL.h
	src/Engine/Options.cpp
	src/Engine/Options.inc.h
	src/Engine/Screen.h
	src/Engine/Surface.cpp
	src/Geoscape/Globe.cpp
	src/Geoscape/GraphsState.cpp
	src/Menu/NewBattleState.cpp
	src/Menu/OptionsVideoState.h
	src/Menu/VideoState.cpp
	src/Mod/Mod.cpp

+++ src/Mod/Armor.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2015 OpenXcom Developers.
+ * Copyright 2010-2016 OpenXcom Developers.
  *
  * This file is part of OpenXcom.
  *
@@ -420,7 +420,7 @@ int Armor::getRankColor(int i) const
 {
 	if ((size_t)i < _rankColor.size())
 	{
-		return _rankColor[i];
+		return _rankColor[(size_t)i];
 	}
 	else
 	{
